Move BaseEntity class (core.base -> core)
